HADOOP-16830 API/Implementation changes for statistics
* tuning standard statistic names
* review javadoc
* can generate a o.a.h.fs.Statistics instance from an IOStatistics instance;
  this can be used to unify statistics collection/reporting.
* Duration tracker to take long as argument
* another duration tracker lambda expression for any InvocationRaisingIOE void -> void.

+++ hadoop-common-project/hadoop-common/src/main/java/org/apache/hadoop/fs/statistics/DurationTrackerFactory.java
@@ -41,7 +41,7 @@ public interface DurationTrackerFactory {
    * operation.
    * @return an object to close after an operation completes.
    */
-  DurationTracker trackDuration(String key, int count);
+  DurationTracker trackDuration(String key, long count);
 
   /**
    * Initiate a duration tracking operation by creating/returning
